# openapi.yaml
openapi: 3.1.0
info:
  title: Samokoder SaaS API
  version: 1.0.0
  description: |
    Полнофункциональный API для платформы Samokoder - AI-powered генератор кода.
    
    ## Аутентификация
    API использует JWT токены с Bearer схемой аутентификации. Токены передаются в заголовке:
    ```
    Authorization: Bearer <access_token>
    ```
    
    ## Безопасность
    - JWT токены с ротацией (access: 15 минут, refresh: 7 дней)
    - Rate limiting на критических эндпоинтах
    - Защита от brute-force атак
    - CORS с белым списком доменов
    - Валидация всех входных данных
    
    ## Версионирование
    API версионируется через префикс `/v1`. Мы гарантируем обратную совместимость в рамках мажорной версии.
    
  contact:
    name: Samokoder Support
    email: support@samokoder.io
  license:
    name: Proprietary
servers:
  - url: https://api.samokoder.io
    description: Production сервер
  - url: https://staging-api.samokoder.io
    description: Staging сервер
  - url: http://localhost:8000
    description: Локальная разработка
...
paths:
...
  /v1/projects/{project_id}/preview/start:
    post:
      tags:
        - preview
      summary: Запустить превью проекта
      description: |
        Запускает dev сервер для превью проекта.
        
        **Реализация:** `api/routers/preview.py:13-25`
        
        **⚠️ ВНИМАНИЕ:** Использует синхронный get_db вместо async get_async_db
      operationId: startPreview
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: Превью запущено
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:
                    type: string
                    format: uri
                    example: "http://localhost:3001"
                  status:
                    type: string
                    example: "running"
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Проект не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - BearerAuth: []
  /v1/projects/{project_id}/preview/stop:
    post:
      tags:
        - preview
      summary: Остановить превью проекта
      description: |
        Останавливает dev сервер превью.
        
        **Реализация:** `api/routers/preview.py:27-35`
        
        **⚠️ TODO:** Реализация не завершена
      operationId: stopPreview
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: Превью остановлено
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Проект не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - BearerAuth: []
  /v1/projects/{project_id}/preview/status:
    get:
      tags:
        - preview
      summary: Получить статус превью
      description: |
        Возвращает текущий статус превью проекта.
        
        **Реализация:** `api/routers/preview.py:37-45`
        
        **⚠️ TODO:** Реализация возвращает хардкоженные данные
      operationId: getPreviewStatus
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: Статус превью
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: object
                    properties:
                      url:
                        type: string
                        format: uri
                      status:
                        type: string
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Проект не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - BearerAuth: []
...
  /v1/usage/token:
    get:
      tags:
        - usage
      summary: Получить статистику использования токенов
      description: |
        Возвращает полную статистику использования токенов пользователя.
        
        **Реализация:** `api/routers/usage.py:11-29`
      operationId: getTokenUsage
      responses:
        '200':
          description: Статистика токенов
          content:
            application/json:
              schema:
                type: object
                properties:
                  usage:
                    type: object
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Ошибка получения статистики
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - BearerAuth: []
  /v1/usage/token/provider/{provider}:
    get:
      tags:
        - usage
      summary: Получить статистику токенов для провайдера
      description: |
        Возвращает статистику использования токенов для конкретного провайдера.
        
        **Реализация:** `api/routers/usage.py:32-64`
      operationId: getProviderUsage
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
          description: Название провайдера
      responses:
        '200':
          description: Статистика провайдера
          content:
            application/json:
              schema:
                type: object
                properties:
                  provider:
                    type: string
                  usage:
                    type: object
                  totals:
                    type: object
                    properties:
                      total_tokens:
                        type: integer
                      total_requests:
                        type: integer
                      models_count:
                        type: integer
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Ошибка получения статистики
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - BearerAuth: []
  /v1/usage/token/provider/{provider}/model/{model}:
    get:
      tags:
        - usage
      summary: Получить статистику токенов для модели
      description: |
        Возвращает статистику использования токенов для конкретной модели провайдера.
        
        **Реализация:** `api/routers/usage.py:67-93`
      operationId: getModelUsage
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
          description: Название провайдера
        - name: model
          in: path
          required: true
          schema:
            type: string
          description: Название модели
      responses:
        '200':
          description: Статистика модели
          content:
            application/json:
              schema:
                type: object
                properties:
                  provider:
                    type: string
                  model:
                    type: string
                  usage:
                    type: object
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Ошибка получения статистики
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - BearerAuth: []
  /v1/usage/token/reset:
    post:
      tags:
        - usage
      summary: Сбросить статистику токенов
      description: |
        Сбрасывает статистику использования токенов.
        Может сбросить все или для конкретного провайдера/модели.
        
        **Реализация:** `api/routers/usage.py:96-132`
        
        **⚠️ ВНИМАНИЕ:** Использует синхронный db.commit() вместо async
      operationId: resetTokenUsage
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                provider:
                  type: string
                model:
                  type: string
      responses:
        '200':
          description: Статистика сброшена
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Ошибка сброса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - BearerAuth: []
  /v1/usage/token/summary:
    get:
      tags:
        - usage
      summary: Получить сводку использования токенов
      description: |
        Возвращает агрегированную сводку использования токенов по всем провайдерам.
        
        **Реализация:** `api/routers/usage.py:135-178`
      operationId: getTokenUsageSummary
      responses:
        '200':
          description: Сводка токенов
          content:
            application/json:
              schema:
                type: object
                properties:
                  totals:
                    type: object
                    properties:
                      providers:
                        type: integer
                      tokens:
                        type: integer
                      requests:
                        type: integer
                  providers:
                    type: object
                    additionalProperties:
                      type: object
                      properties:
                        tokens:
                          type: integer
                        requests:
                          type: integer
                        models:
                          type: integer
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Ошибка получения сводки
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - BearerAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT токен в формате: Bearer <token>
        
        Получить токен можно через `/v1/auth/login` или `/v1/auth/register`
  parameters:
    ProjectId:
      name: project_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Уникальный идентификатор проекта
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
          description: Описание ошибки
        error_code:
          type: string
          description: Код ошибки (опционально)
        timestamp:
          type: string
          format: date-time
          description: Время возникновения ошибки
...
security:
  - BearerAuth: []
...